port done until at least compiling

+++ src/game/menuscreen.cpp
@@ -1,12 +1,10 @@
-#include <allegro5/allegro.h>
-#include <allegro5/allegro_primitives.h>
-#include <alligator/input/input.h>
+#include <aether/aether.h>
 #include <iostream>
 
 #include "menuscreen.h"
 #include "phackmangame.h"
 #include "assets.h"
-#include <alligator/util/matrix.h>
+
 //#include "../debug/mapsoliddebug.h"
 
 MenuScreen::MenuScreen( PhackmanGame* g )
@@ -20,7 +18,7 @@ void MenuScreen::show()
 
 }
 
-void MenuScreen::update(double delta)
+void MenuScreen::update(uint64_t delta)
 {
 
 }
